Unify build, builder, and init
